requirements.txt
# CSAO Rail Recommendation System - Dependencies
# Python 3.10+

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.2.0

# Config & utils
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# Optional: tuning & evaluation
optuna>=3.5.0

# Dev / testing
pytest>=7.0.0
httpx>=0.26.0
